+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # v2.12.2
 
-
+* added `HashableMixin`
+* `HashableWrapper` now supports __eq__
+* added `build_structure`