+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 * added `HashableMixin`
 * `HashableWrapper` now supports __eq__
 * added `build_structure`
+* added `x.type`
+* added `replace_argument_if`