+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-# v2.12.3
-
-* added `tags_factory` to `trace_function`
+# v2.12.4